+++ tower-http/CHANGELOG.md
@@ -26,13 +26,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **trace:** Correctly identify gRPC requests in default `on_response` callback ([#278])
 - **cors:** Panic if a wildcard (`*`) is passed to `AllowOrigin::list`. Use
   `AllowOrigin::any()` instead ([#285])
+- **serve_dir:** Call the fallback on non-uft8 request paths ([#310])
 
+[#275]: https://github.com/tower-rs/tower-http/pull/275
 [#278]: https://github.com/tower-rs/tower-http/pull/278
 [#285]: https://github.com/tower-rs/tower-http/pull/285
-[#275]: https://github.com/tower-rs/tower-http/pull/275
 [#289]: https://github.com/tower-rs/tower-http/pull/289
 [#299]: https://github.com/tower-rs/tower-http/pull/299
 [#303]: https://github.com/tower-rs/tower-http/pull/303
+[#310]: https://github.com/tower-rs/tower-http/pull/310
 
 # 0.3.4 (June 06, 2022)